restructure project, move all source code under core folder
