+++ core/layers/__init__.py
@@ -1,1 +1,2 @@
 from .layer import Layer
+from .input import Input
